refactor: 조회 전용 repository 정리 #222
