+++ server/src/main/java/com/hobbyt/domain/mypage/repository/MyPageRepository.java
@@ -1,11 +1,13 @@
-package com.hobbyt.domain.order.repository;
+package com.hobbyt.domain.mypage.repository;
 
 import org.springframework.data.domain.Pageable;
 
 import com.hobbyt.domain.mypage.dto.PageDto;
 import com.hobbyt.domain.mypage.dto.response.OrderDetails;
 
-public interface CustomOrderRepository {
+public interface MyPageRepository {
+	PageDto getOrderedProductsByMemberEmail(String email, Pageable pageable);
+
 	OrderDetails findOrderDetailsByOrderId(Long orderId);
 
 	PageDto findMyOrdersByEmail(String email, Pageable pageable);
