+++ server/src/main/java/com/hobbyt/domain/mypage/service/MyPageService.java
@@ -6,16 +6,14 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.hobbyt.domain.member.service.MemberService;
 import com.hobbyt.domain.mypage.dto.OrderedProductInfo;
 import com.hobbyt.domain.mypage.dto.PageDto;
 import com.hobbyt.domain.mypage.dto.response.OrderDetails;
 import com.hobbyt.domain.mypage.dto.response.PageResponse;
+import com.hobbyt.domain.mypage.repository.MyPageRepository;
 import com.hobbyt.domain.order.entity.Order;
 import com.hobbyt.domain.order.entity.OrderStatus;
-import com.hobbyt.domain.order.repository.OrderRepository;
 import com.hobbyt.domain.order.service.OrderService;
-import com.hobbyt.domain.sale.repository.ProductRepository;
 import com.hobbyt.global.entity.Account;
 import com.hobbyt.global.entity.Address;
 
@@ -25,13 +23,11 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
 public class MyPageService {
-	private final MemberService memberService;
 	private final OrderService orderService;
-	private final ProductRepository productRepository;
-	private final OrderRepository orderRepository;
+	private final MyPageRepository myPageRepository;
 
 	public PageResponse getOrderedProducts(String email, Pageable pageable) {
-		PageDto pageDto = productRepository.getOrderedProductsByMemberEmail(email, pageable);
+		PageDto pageDto = myPageRepository.getOrderedProductsByMemberEmail(email, pageable);
 
 		Page<OrderedProductInfo> page = new PageImpl<>(pageDto.getContent(), pageable, pageDto.getTotal());
 
@@ -46,20 +42,20 @@ public class MyPageService {
 	}
 
 	public OrderDetails getOrderDetails(Long orderId) {
-		OrderDetails orderDetails = orderRepository.findOrderDetailsByOrderId(orderId);
+		OrderDetails orderDetails = myPageRepository.findOrderDetailsByOrderId(orderId);
 		return orderDetails;
 	}
 
-	// 내가한 주문
+	// 내가한 주문 TODO test
 	public PageResponse getMyOrders(String email, Pageable pageable) {
-		PageDto pageDto = orderRepository.findMyOrdersByEmail(email, pageable);
+		PageDto pageDto = myPageRepository.findMyOrdersByEmail(email, pageable);
 		Page<OrderedProductInfo> page = new PageImpl<>(pageDto.getContent(), pageable, pageDto.getTotal());
 		return PageResponse.of(page);
 	}
 
 	// 내가 올린 판매글에 대한 주문
 	public PageResponse getOrders(String email, Pageable pageable) {
-		PageDto pageDto = orderRepository.findOrdersByEmail(email, pageable);
+		PageDto pageDto = myPageRepository.findOrdersByEmail(email, pageable);
 		Page<OrderedProductInfo> page = new PageImpl<>(pageDto.getContent(), pageable, pageDto.getTotal());
 		return PageResponse.of(page);
 	}
