+++ server/src/main/java/com/hobbyt/domain/order/repository/OrderRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.hobbyt.domain.order.entity.Order;
 
-public interface OrderRepository extends JpaRepository<Order, Long>, CustomOrderRepository {
+public interface OrderRepository extends JpaRepository<Order, Long> {
 }
