+++ server/src/main/java/com/hobbyt/domain/sale/repository/CustomProductRepositoryImpl.java
@@ -1,6 +1,5 @@
 package com.hobbyt.domain.sale.repository;
 
-import static com.hobbyt.domain.member.entity.QMember.*;
 import static com.hobbyt.domain.order.entity.QOrderItem.*;
 import static com.hobbyt.domain.sale.entity.QProduct.*;
 import static com.hobbyt.domain.sale.entity.QSale.*;
@@ -9,11 +8,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import org.springframework.data.domain.Pageable;
-
-import com.hobbyt.domain.mypage.dto.OrderedProductInfo;
-import com.hobbyt.domain.mypage.dto.PageDto;
-import com.querydsl.core.types.Projections;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 
 import lombok.RequiredArgsConstructor;
@@ -38,43 +32,5 @@ public class CustomProductRepositoryImpl implements CustomProductRepository {
 				.fetchFirst() != null)
 		);
 		return result;
-	}
-
-	@Override
-	public PageDto getOrderedProductsByMemberEmail(String email, Pageable pageable) {
-		List<OrderedProductInfo> content = queryFactory.select(
-				Projections.constructor(OrderedProductInfo.class,
-					sale.id,
-					sale.writer.id,
-					product.name,
-					sale.period,
-					sale.isAlwaysOnSale,
-					sale.isDeleted,
-					product.salesVolume,
-					product.createdAt
-				))
-			.from(sale)
-			.join(sale.products, product)
-			.join(sale.writer, member)
-			.where(member.email.eq(email))
-			.orderBy(sale.id.desc(), product.id.desc())
-			.offset(pageable.getOffset())
-			.limit(pageable.getPageSize())
-			.fetch();
-
-		Long total = getOrderedProductsCount(email);
-
-		return new PageDto(content, total);
-	}
-
-	private Long getOrderedProductsCount(String email) {
-		return queryFactory.select(
-				product.count()
-			)
-			.from(sale)
-			.join(sale.products, product)
-			.join(sale.writer, member)
-			.where(member.email.eq(email))
-			.fetchOne();
 	}
 }
